[TableGen] Spit GlobalISelCombinerEmitter into multiple files
Split MatchDataInfo, CXXPredicates and the Pattern hierarchy into their own files.

This should help with maintenance a bit, and make the API easier to navigate.

I also hope this encourages a bit more experimentation with MIR patterns, e.g. I'd like to try getting them in ISel at some point.
